make optimzier.step as hook
