+++ cvpods/engine/__init__.py
@@ -3,7 +3,8 @@
 from .hooks import *
 from .launch import *
 from .predictor import *
-from .setup import *
+from .base_runner import *
 from .runner import *
+from .setup import *
 
 __all__ = [k for k in globals().keys() if not k.startswith("_")]
