+++ cvpods/modeling/__init__.py
@@ -5,11 +5,11 @@ from cvpods.layers import ShapeSpec
 
 # from .anchor_generator import build_anchor_generator
 from .backbone import FPN, Backbone, ResNet, ResNetBlockBase, build_resnet_backbone, make_stage
+from .matcher import Matcher
 from .meta_arch import GeneralizedRCNN, PanopticFPN, ProposalNetwork, RetinaNet, SemanticSegmentor
 from .postprocessing import detector_postprocess
 from .roi_heads import ROIHeads, StandardROIHeads
 from .test_time_augmentation import DatasetMapperTTA, GeneralizedRCNNWithTTA, TTAWarper
-from .matcher import Matcher
 
 _EXCLUDE = {"torch", "ShapeSpec"}
 __all__ = [k for k in globals().keys() if k not in _EXCLUDE and not k.startswith("_")]
